moved files and re-added dashboard charts
