--- app/components/widgets/charts/AssetDistribution/AssetDistribution.tsx
@@ -1,24 +1,0 @@
-import { DonutChart } from "@tremor/react"
-export default function AssetDistribution() {
-  const chartData = [
-    {
-      label: "Cash",
-      value: 5000,
-    },
-    {
-      label: "Super",
-      value: 18000,
-    },
-  ]
-  return (
-    <DonutChart
-      className="h-full"
-      data={chartData}
-      category="value"
-      index="label"
-      showAnimation
-      //valueFormatter={valueFormatter}
-      colors={["slate", "violet", "indigo", "rose", "cyan", "amber"]}
-    />
-  )
-}
