--- app/components/widgets/charts/AssetDistribution/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./AssetDistribution"
