--- app/components/widgets/charts/NetWorth/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./NetWorth"
